+++ CHANGELOG.md
@@ -8,10 +8,14 @@
 ### Added
 - Console log messages when building with the VS Code extension
 - `.print-float-right` and `.print-float-left` attributes to float only in print layouts
+- VS Code E+ View now accurately shows order of pages and groups
 
 ### Changed
 - Swapped page footer orientation between odd and even pages 
 - Decreased font size of figure captions
+
+### Fixed
+- Newline handling in Trait/Action descriptions for monsters
 
 ## 1.0.15